+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog] and this project adheres to
 - `meta` module with Ruby's metadata
 - `From<()>` implementation for `AnyObject`
   - Allows for not returning anything in `def_method[_unchecked]!`
+- `InstrSeq::eval_unchecked`
 
 ### Changed
 - Where `Float` and `Integer` are located; both now reside in a `num` module